[리팩토링] Restaurant 관련 컴포넌트의 구조 변경
